+++ packages/core/src/runtime/App.tsx
@@ -10,10 +10,8 @@ import { HelmetProvider } from 'react-helmet-async';
 import React, { useContext, useLayoutEffect } from 'react';
 import { Header, PageData, cleanUrl, MDX_REGEXP } from '@rspress/shared';
 import globalComponents from 'virtual-global-components';
+import Theme from '@theme';
 import 'virtual-global-styles';
-
-// eslint-disable-next-line import/no-commonjs
-const { default: Theme } = require('@theme');
 
 export enum QueryStatus {
   Show = '1',
@@ -39,7 +37,9 @@ export async function initPageData(routePath: string): Promise<PageData> {
     const extractPageInfo = siteData.pages.find(page => {
       const normalize = (p: string) =>
         // compat the path that has no / suffix and ignore case
-        p.replace(/\/$/, '').toLowerCase();
+        p
+          .replace(/\/$/, '')
+          .toLowerCase();
       return isEqualPath(normalize(page.routePath), normalize(routePath));
     });
 
@@ -68,24 +68,24 @@ export async function initPageData(routePath: string): Promise<PageData> {
         toc,
       },
     };
-  } else {
-    // 404 Page
-    return {
-      siteData,
-      page: {
-        pagePath: '',
-        pageType: '404',
-        routePath: '/404',
-        lang: siteData.lang || '',
-        frontmatter: {},
-        title: '404',
-        toc: [],
-        version: '',
-        _filepath: '',
-        _relativePath: '',
-      },
-    };
   }
+
+  // 404 Page
+  return {
+    siteData,
+    page: {
+      pagePath: '',
+      pageType: '404',
+      routePath: '/404',
+      lang: siteData.lang || '',
+      frontmatter: {},
+      title: '404',
+      toc: [],
+      version: '',
+      _filepath: '',
+      _relativePath: '',
+    },
+  };
 }
 
 export function App({ helmetContext }: { helmetContext?: object }) {
@@ -125,12 +125,12 @@ export function App({ helmetContext }: { helmetContext?: object }) {
                 key: index,
                 ...props,
               });
-            } else {
-              return React.createElement(componentInfo, {
-                // eslint-disable-next-line react/no-array-index-key
-                key: index,
-              });
             }
+
+            return React.createElement(componentInfo, {
+              // eslint-disable-next-line react/no-array-index-key
+              key: index,
+            });
           })
       }
     </HelmetProvider>