+++ packages/plugin-preview/src/utils.ts
@@ -7,10 +7,10 @@ export const toValidVarName = (str: string): string => {
   // Check if the string is a valid variable name
   if (/^[a-zA-Z_$][0-9a-zA-Z_$]*$/.test(str)) {
     return str; // If it is a valid variable name, return the original string
-  } else {
-    // If it is not a valid variable name, convert it to a valid variable name
-    return str.replace(/[^0-9a-zA-Z_$]/g, '_').replace(/^([0-9])/, '_$1');
   }
+
+  // If it is not a valid variable name, convert it to a valid variable name
+  return str.replace(/[^0-9a-zA-Z_$]/g, '_').replace(/^([0-9])/, '_$1');
 };
 
 export const generateId = (pageName: string, index: number) => {