+++ packages/theme-default/src/components/Search/SearchPanel.tsx
@@ -44,15 +44,37 @@ export function SearchPanel({ focused, setFocused }: SearchPanelProps) {
   const [initing, setIniting] = useState(true);
   const [currentSuggestionIndex, setCurrentSuggestionIndex] = useState(0);
   const pageSearcherRef = useRef<PageSearcher | null>(null);
-  const searchResultRef = useRef<HTMLInputElement | null>(null);
+  const searchResultRef = useRef(null);
+  const searchResultTabRef = useRef(null);
 
   // only scroll after keydown arrow up and arrow down.
   const [canScroll, setCanScroll] = useState(false);
-  const scrollTo = (top: number) => {
+  const scrollTo = (offsetTop: number, offsetHeight: number) => {
     if (canScroll) {
-      searchResultRef?.current?.scrollTo({
-        top,
-      });
+      // Down
+      // 50 = 20(modal margin) + 40(input height) - 10(item margin)
+      // -10 = 50(following) - 50(tab title) - 10(item margin)
+      const scrollDown =
+        offsetTop +
+        offsetHeight -
+        searchResultRef?.current?.offsetHeight -
+        (searchResult.length === 1 ? 50 : -10);
+      if (scrollDown > searchResultRef?.current?.scrollTop) {
+        searchResultRef?.current?.scrollTo({
+          top: scrollDown,
+        });
+      }
+
+      // Up
+      // 70 = 20(modal margin) + 40(input height) + 10(item margin)
+      // 10 = 70(following) - 50(tab title) - 10(item margin)
+      const scrollUp =
+        searchResult.length === 1 ? offsetTop - 70 : offsetTop - 10;
+      if (scrollUp < searchResultRef?.current?.scrollTop) {
+        searchResultRef?.current?.scrollTo({
+          top: scrollUp,
+        });
+      }
     }
   };
   const {
@@ -257,7 +279,11 @@ export function SearchPanel({ focused, setFocused }: SearchPanelProps) {
       if (currentSearchResult.length === 0) {
         return <NoSearchResult query={query} />;
       }
-      return <div>{renderSearchResultItem(currentSearchResult)}</div>;
+      return (
+        <div ref={searchResultTabRef}>
+          {renderSearchResultItem(currentSearchResult)}
+        </div>
+      );
     }
 
     const tabValues = result.map(item => {
@@ -281,6 +307,8 @@ export function SearchPanel({ focused, setFocused }: SearchPanelProps) {
           setCurrentSuggestionIndex(0);
           setCurrentRenderType(result[index].renderType);
         }}
+        // @ts-ignore
+        ref={searchResultTabRef}
       >
         {result.map(item => (
           <Tab key={item.group}>
@@ -326,7 +354,6 @@ export function SearchPanel({ focused, setFocused }: SearchPanelProps) {
                   return (
                     <SuggestItem
                       key={`${suggestion.title}-${suggestionIndex}`}
-                      suggestionIndex={suggestionIndex}
                       suggestion={suggestion}
                       isCurrent={suggestionIndex === currentSuggestionIndex}
                       setCurrentSuggestionIndex={() => {