+++ packages/theme-default/src/components/Search/SuggestItem.tsx
@@ -22,25 +22,22 @@ export function SuggestItem({
   setCurrentSuggestionIndex,
   inCurrentDocIndex,
   scrollTo,
-  suggestionIndex,
 }: {
   suggestion: DefaultMatchResultItem;
   closeSearch: () => void;
   isCurrent: boolean;
   setCurrentSuggestionIndex: () => void;
   inCurrentDocIndex: boolean;
-  scrollTo: (top: number) => void;
-  suggestionIndex: number;
+  scrollTo: (top: number, height: number) => void;
 }) {
   const HitIcon = ICON_MAP[suggestion.type];
   const link =
     inCurrentDocIndex && !isProduction()
       ? removeDomain(suggestion.link)
       : suggestion.link;
-  const selfRef = useRef<HTMLInputElement | null>(null);
+  const selfRef = useRef(null);
   if (isCurrent) {
-    // TODO: only scroll when unvisible
-    scrollTo(selfRef?.current?.offsetTop - 80);
+    scrollTo(selfRef?.current?.offsetTop, selfRef?.current?.offsetHeight);
   }
 
   const getHighlightedFragments = (
@@ -80,10 +77,11 @@ export function SuggestItem({
           {getHighlightedFragments(header, highlightInfoList)}
         </div>
       );
-    } else {
-      return <div className="font-medium">{suggestion.header}</div>;
     }
+
+    return <div className="font-medium">{suggestion.header}</div>;
   };
+
   const renderStatementMatch = () => {
     if (suggestion.type !== 'content') {
       return <div></div>;