+++ packages/theme-default/src/layout/DocLayout/docComponents/link.tsx
@@ -1,7 +1,7 @@
 import { ComponentProps } from 'react';
 import styles from './index.module.scss';
-import { Link } from '#theme/components/Link';
-import { usePathUtils } from '#theme/logic';
+import { Link } from '@theme';
+import { usePathUtils } from '../../../logic';
 
 export const A = (props: ComponentProps<'a'>) => {
   const { href = '', className = '' } = props;
@@ -10,13 +10,13 @@ export const A = (props: ComponentProps<'a'>) => {
 
   if (hasHeaderAnchor || href.startsWith('#')) {
     return <a {...props} className={`${styles.link} ${className}`} />;
-  } else {
-    return (
-      <Link
-        {...props}
-        className={`${className} ${styles.link} ${styles['inline-link']}`}
-        href={normalizeLinkHref(href)}
-      />
-    );
   }
+
+  return (
+    <Link
+      {...props}
+      className={`${className} ${styles.link} ${styles['inline-link']}`}
+      href={normalizeLinkHref(href)}
+    />
+  );
 };