+++ packages/document/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @rspress/docs
 
+## 1.16.0
+
+## 1.15.1
+
+## 1.15.0
+
+### Patch Changes
+
+- 2a10e2d: feat: support showing line numbers in shiki plugin
+
 ## 1.14.0
 
 ### Patch Changes