+++ packages/plugin-playground/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @rspress/plugin-playground
 
+## 1.16.0
+
+### Patch Changes
+
+- Updated dependencies [4cf049c]
+- Updated dependencies [516c8cc]
+  - @rspress/shared@1.16.0
+
+## 1.15.1
+
+### Patch Changes
+
+- @rspress/shared@1.15.1
+
+## 1.15.0
+
+### Patch Changes
+
+- @rspress/shared@1.15.0
+
 ## 1.14.0
 
 ### Patch Changes